backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymysql==1.1.1
cryptography==42.0.8
gunicorn==22.0.0
